+++ core/src/main/scala/doobie/util/connection.scala
@@ -7,8 +7,10 @@ import scalaz.{ Monad, Catchable }
 
 import java.sql.Connection
 
+/** Typeclass instances for `java.sql.Connection`. */
 object connection {
 
+  /** @group Typeclass Instances */
   implicit def connectionConnector[M[_]: Monad: Capture: Catchable]: Connector[M, Connection] =
     Connector.instance(Monad[M].point(_))
