+++ core/src/main/scala/doobie/util/transactor.scala
@@ -3,7 +3,7 @@ package doobie.util
 import doobie.util.capture.Capture
 import doobie.free.connection.ConnectionIO
 import doobie.util.liftxa.LiftXA
-import doobie.util.connector.{ Connector, ConnectionIOConnectorOps }
+import doobie.util.connector.Connector
 import doobie.util.yolo.Yolo
 import doobie.hi.connection.ProcessConnectionIOOps
 
@@ -51,21 +51,5 @@ object transactor {
       liftXA.get(t).safeP(pa)
 
   }
-
-  /** Syntax that adds `Transactor` operations to `ConnectionIO`. */
-  implicit class ConnectionIOTransactorOps[A](ma: ConnectionIO[A]) {
-
-    final class Helper[M[_]] {
-      def apply[T](t: T)(implicit ev1: Transactor[M, T]): M[A] =
-        Connector.trans(t).apply(Transactor.safe(t)(ma))
-    }
-
-    /** 
-     * Interpret this program configured with transaction logic as defined by the `Transactor`'s
-     * `LiftXA`, into effect-capturing target monad `M`, running on a dedicated `Connection`. 
-     */
-    def transact[M[_]] = new Helper[M]
-
-  }
-
+  
 }
